+++ src/Symfony/Component/Console/Resources/completion.bash
@@ -18,7 +18,8 @@ _sf_{{ COMMAND_NAME }}() {
 
     # Use newline as only separator to allow space in completion values
     IFS=$'\n'
-    local sf_cmd="${COMP_WORDS[0]}"
+    # _SF_CMD allows Symfony CLI to tell us to use a different command to run the console
+    local sf_cmd="${_SF_CMD:-${COMP_WORDS[0]}}"
 
     # for an alias, get the real script behind it
     sf_cmd_type=$(type -t $sf_cmd)